+++ chaosazure/machine/actions.py
@@ -10,7 +10,8 @@ from chaosazure import auth
 from chaosazure.machine.constants import RES_TYPE_VM
 from chaosazure.rgraph.resource_graph import fetch_resources
 
-__all__ = ["delete_machine", "stop_machine", "restart_machine"]
+__all__ = ["delete_machine", "stop_machine", "restart_machine",
+           "start_machine"]
 
 
 def delete_machine(filter: str = None,
@@ -34,10 +35,10 @@ def delete_machine(filter: str = None,
         "Start delete_machine: configuration='{}', filter='{}'".format(
             configuration, filter))
 
-    choice = choose_machine_at_random(filter, configuration, secrets)
+    choice = __fetch_machine_at_random(filter, configuration, secrets)
 
     logger.debug("Deleting machine: {}".format(choice['name']))
-    client = init_client(secrets, configuration)
+    client = __init_client(secrets, configuration)
     client.virtual_machines.delete(choice['resourceGroup'], choice['name'])
 
 
@@ -59,10 +60,10 @@ def stop_machine(filter: str = None,
         "Start stop_machine: configuration='{}', filter='{}'".format(
             configuration, filter))
 
-    choice = choose_machine_at_random(filter, configuration, secrets)
+    choice = __fetch_machine_at_random(filter, configuration, secrets)
 
     logger.debug("Stopping machine: {}".format(choice['name']))
-    client = init_client(secrets, configuration)
+    client = __init_client(secrets, configuration)
     client.virtual_machines.power_off(choice['resourceGroup'], choice['name'])
 
 
@@ -84,30 +85,82 @@ def restart_machine(filter: str = None,
         "Start restart_machine: configuration='{}', filter='{}'".format(
             configuration, filter))
 
-    choice = choose_machine_at_random(filter, configuration, secrets)
+    choice = __fetch_machine_at_random(filter, configuration, secrets)
 
     logger.debug("Restarting machine: {}".format(choice['name']))
-    client = init_client(secrets, configuration)
+    client = __init_client(secrets, configuration)
     client.virtual_machines.restart(choice['resourceGroup'], choice['name'])
+
+
+def start_machine(filter: str = None,
+                  configuration: Configuration = None,
+                  secrets: Secrets = None):
+    """
+    Start a virtual machine that is in a stopped state.
+
+    Parameters
+    ----------
+    filter : str
+        Filter the virtual machines. If the filter is omitted all stopped
+        machines in the subscription will be started again.
+        Filtering example:
+        'where resourceGroup=="myresourcegroup" and name="myresourcename"'
+    """
+
+    logger.debug(
+        "Start start_machine: configuration='{}', filter='{}'".format(
+            configuration, filter))
+
+    machines = __fetch_machines(filter, configuration, secrets)
+    client = __init_client(secrets, configuration)
+    stopped_machines = __fetch_stopped_machines(client, machines)
+    __start_stopped_machines(client, stopped_machines)
 
 
 ###############################################################################
 # Private helper functions
 ###############################################################################
-def choose_machine_at_random(filter, configuration, secrets):
+def __start_stopped_machines(client, stopped_machines):
+    for machine in stopped_machines:
+        logger.debug("Starting machine: {}".format(machine['name']))
+        client.virtual_machines.start(machine['resourceGroup'],
+                                      machine['name'])
+
+
+def __fetch_stopped_machines(client, machines):
+    stopped_machines = []
+    for m in machines:
+        i = client.virtual_machines.instance_view(m['resourceGroup'],
+                                                  m['name'])
+        for s in i.statuses:
+            status = s.code.lower().split('/')
+            if status[0] == 'powerstate' and (
+                    status[1] == 'deallocated' or status[1] == 'stopped'):
+                stopped_machines.append(m)
+                logger.debug("Found stopped machine: {}".format(m['name']))
+    return stopped_machines
+
+
+def __fetch_machines(filter, configuration, secrets):
     machines = fetch_resources(filter, RES_TYPE_VM, secrets, configuration)
     if not machines:
         logger.warning("No virtual machines found")
         raise FailedActivity("No virtual machines found")
     else:
         logger.debug(
-            "Found virtual machines: {}".format(
+            "Fetched virtual machines: {}".format(
                 [x['name'] for x in machines]))
+    return machines
+
+
+def __fetch_machine_at_random(filter, configuration, secrets):
+    machines = __fetch_machines(
+        filter, configuration=configuration, secrets=secrets)
     choice = random.choice(machines)
     return choice
 
 
-def init_client(secrets, configuration):
+def __init_client(secrets, configuration):
     with auth(secrets) as cred:
         subscription_id = configuration['azure']['subscription_id']
         client = ComputeManagementClient(